pyproject.toml
[tool.poetry]
name = "authz-demo"
version = "0.1.0"
description = ""
authors = ["NBadriddin <nbadriddin70@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^6.0"
djangorestframework = "^3.16.1"
django-cors-headers = "^4.9.0"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
